arch: update structure .py file
