+++ data_manipulation_service/src/core/dataset_module/filesystem/__init__.py
@@ -1,0 +1,4 @@
+from .am import ArchiveManager
+from .fsm import FileSystemManager
+
+__all__ = ['FileSystemManager', 'ArchiveManager']
